refactor: availableBalances component
