+++ apps/portal/src/components/widgets/staking/slpx/StakeProviders.tsx
@@ -1,46 +1,26 @@
-import { selectedBalancesState, selectedCurrencyState } from '../../../../domains/balances'
+import { selectedBalancesState } from '../../../../domains/balances'
 import { ChainProvider } from '../../../../domains/chains'
 import { slpxPairsState, type SlpxPair } from '../../../../domains/staking/slpx'
 import StakeProvider from '../../../recipes/StakeProvider'
 import AnimatedFiatNumber from '../../AnimatedFiatNumber'
 import ErrorBoundary from '../../ErrorBoundary'
-import RedactableBalance from '../../RedactableBalance'
 import ErrorBoundaryFallback from '../ErrorBoundaryFallback'
 import Apr from './Apr'
+import { AvailableBalance, useAvailableBalance } from './AvailableBalances'
 import UnlockDuration from './UnlockDuration'
-import { Decimal } from '@talismn/math'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import BigNumber from 'bignumber.js'
 import { useMemo } from 'react'
 import { Link } from 'react-router-dom'
-import { useRecoilValue, waitForAll } from 'recoil'
+import { useRecoilValue } from 'recoil'
 import { useConfig } from 'wagmi'
 import { getTokenQueryOptions } from 'wagmi/query'
-
-const useAvailableBalance = (slpxPair: SlpxPair) => {
-  const [balances, currency] = useRecoilValue(waitForAll([selectedBalancesState, selectedCurrencyState]))
-  const nativeBalance = balances.find(x => x.token?.symbol.toLowerCase() === slpxPair.nativeToken.symbol.toLowerCase())
-
-  return useMemo(
-    () => ({
-      amount: Decimal.fromPlanck(nativeBalance.sum.planck.transferable ?? 0n, nativeBalance.each.at(0)?.decimals ?? 0, {
-        currency: slpxPair.nativeToken.symbol,
-      }).toLocaleString(),
-      fiatAmount: nativeBalance.sum.fiat(currency).total,
-    }),
-    [currency, nativeBalance.each, nativeBalance.sum, slpxPair.nativeToken.symbol]
-  )
-}
-
-const AvailableBalance = (props: { slpxPair: SlpxPair }) => (
-  <RedactableBalance>{useAvailableBalance(props.slpxPair).amount}</RedactableBalance>
-)
 
 const AvailableFiatBalance = (props: { slpxPair: SlpxPair }) => (
   <AnimatedFiatNumber end={useAvailableBalance(props.slpxPair).fiatAmount} />
 )
 
-const StakePercentage = (props: { slpxPair: SlpxPair }) => {
+export const StakePercentage = (props: { slpxPair: SlpxPair }) => {
   const balances = useRecoilValue(selectedBalancesState)
 
   const config = useConfig()
